feat(form-refactor): moved some logic from Input component and transformed into a hook, added an .md file to explain it's usage
