+++ packages/lumina-core/src/components/form-components-RHF/index.scss
@@ -4,5 +4,5 @@
 @import './switch/styles';
 @import './textarea/styles';
 @import './slider/styles';
-@import './utils/help/styles';
-@import './utils/label/styles';
+@import './utils/utility-components/help/styles';
+@import './utils/utility-components/label/styles';
